Client v2
FolderStatus, Conflicts checks, etc.

Before refactor

+++ ServerFileSync/ServerFileSync/Controllers/FileTransferController.cs
@@ -1,4 +1,5 @@
 ﻿using ServerFileSync.Interfaces;
+using ServerFileSync.Models;
 using System;
 using System.Collections.Generic;
 using System.Configuration;
@@ -17,7 +18,7 @@ namespace ServerFileSync.Controllers
     public class FileTransferController : ApiController
     {
         #region Class vars
-        private string _root;
+        private string _syncFolder;
         private IFileManager _fileManager;
         private IFileNotifier _hubWrapper;
         #endregion Class vars
@@ -28,8 +29,8 @@ namespace ServerFileSync.Controllers
         /// </summary>
         public FileTransferController()
         {
-            _root = ConfigurationManager.AppSettings["SyncFolder"].ToString();
-            _fileManager = new FileSystemFileManager(_root);
+            _syncFolder = ConfigurationManager.AppSettings["SyncFolder"].ToString();
+            _fileManager = new FileSystemFileManager(_syncFolder);
             _hubWrapper = FileSyncHubWrapper.Instance;
         }
 
@@ -38,12 +39,12 @@ namespace ServerFileSync.Controllers
         /// </summary>
         /// <param name="fileManager"></param>
         /// <param name="hubWrapper"></param>
-        /// <param name="root"></param>
-        public FileTransferController(IFileManager fileManager, IFileNotifier hubWrapper, string root)
+        /// <param name="syncFolder"></param>
+        public FileTransferController(IFileManager fileManager, IFileNotifier hubWrapper, string syncFolder)
         {
             _fileManager = fileManager;
             _hubWrapper = hubWrapper;
-            _root = root;
+            _syncFolder = syncFolder;
         }
         #endregion Constructors
 
@@ -55,9 +56,9 @@ namespace ServerFileSync.Controllers
         /// Generates SignalR NotifyNewFile on success
         /// </summary>
         /// <param name="fileName">Original name of the file</param>
-        /// <returns></returns>
+        /// <returns>HttpResponseMessage</returns>
         [HttpPost]
-        public HttpResponseMessage ConfirmSave(string fileName)
+        public HttpResponseMessage ConfirmUpload(string fileName)
         {
             Guid tempGuid;
             try
@@ -71,7 +72,7 @@ namespace ServerFileSync.Controllers
             }
 
             if (!_fileManager.ExistsTemp(fileName,tempGuid))
-                return new HttpResponseMessage(HttpStatusCode.BadRequest);
+                return new HttpResponseMessage(HttpStatusCode.NotFound);
 
             try
             {
@@ -122,10 +123,23 @@ namespace ServerFileSync.Controllers
         }
 
         /// <summary>
-        /// Uploads a file and saves it with a temporary name with a generated Guid
-        /// If the uploaded file exists with the same Hash, nothing is done
+        /// Uploads a file
+        /// If the file to overwrite is different than the expected, keeps the uploaded file with a temporary name generated with a Guid. Need to complete the operation calling ConfirmUpload or DeleteTemp
+        /// The generated Guid is returned as a String Content in the HttpResponse
+        /// Expected ordered Request Content:
+        /// Filename as string
+        /// Previous file hash as string
+        /// File content as byte array
         /// </summary>
-        /// <returns>The generated Guid as a String Content in the HttpResponse</returns>
+        /// <returns>
+        /// HttpResponseMessage
+        /// The generated Guid as String Content in the HttpResponse
+        /// HttpResponse HttpStatusCode:
+        /// OK: No file existed with that name, Upload confirmed. Generates SignalR NotifyNewFile
+        /// Accepted: The file to overwrite was the expected one, Upload confirmed. Generates SignalR NotifyNewFile
+        /// NotModified: A file with same name and hash existed, no need to Upload
+        /// Ambiguous: A file with same name exists, but different hash than the expected and the uploaded one. File was kept with a temporary name. Guid returned. Need to call ConfirmUpload to overwrite, or DeleteTemp to cancel
+        /// </returns>
         [HttpPost]
         public async Task<HttpResponseMessage> Upload()
         {
@@ -140,30 +154,42 @@ namespace ServerFileSync.Controllers
             try
             {
                 string fileName = await getStringDataFromRequest(multiContents,0);
+                string previousHash = await getStringDataFromRequest(multiContents, 1);
                 byte[] fileBytes = await getFileBytesFromRequest(multiContents);
                 
-                Guid tempGuid = _fileManager.Save(fileName, fileBytes);
-
                 HttpResponseMessage myResponse = new HttpResponseMessage();
-                myResponse.Content = new StringContent(tempGuid.ToString());
-                myResponse.Content.Headers.ContentType = new MediaTypeHeaderValue("text/plain");
 
                 //Check original file existance at the end of the temporal copy
                 if (_fileManager.Exists(fileName))
                 {
-                    //If the existing File has the same Hash, nothing needs to be done
+                    //If the existing File has the same Hash, nothing needs to be done, the updated file was already uploaded
                     if (_fileManager.SameHash(fileName, fileBytes))
-                    {
-                        _fileManager.DeleteTemp(fileName, tempGuid);
                         return new HttpResponseMessage(HttpStatusCode.NotModified);
-                    }
 
-                    myResponse.StatusCode = HttpStatusCode.Ambiguous;
+                    string CRC = _fileManager.GetHash(fileName);
+
+                    //If the existing File has the client expected Hash, automatically ConfirmUpload
+                    if (CRC.Equals(previousHash))
+                    {
+                        _fileManager.Save(fileName, fileBytes, false);
+                        _hubWrapper.NotifyNewFile(fileName, CRC);
+                        myResponse.StatusCode = HttpStatusCode.Accepted;
+                        return myResponse;
+                    }
                 }
-                else
+                else //File is New, automatically ConfirmUpload
                 {
+                    _fileManager.Save(fileName, fileBytes, false);
+                    _hubWrapper.NotifyNewFile(fileName, _fileManager.GetHash(fileName));
                     myResponse.StatusCode = HttpStatusCode.OK;
+                    return myResponse;
                 }
+
+                //File exists on Server, but Hash is different than expected. User must decide and call ConfirmUpload or DeleteTemp. Temp Guid returned.
+                myResponse.StatusCode = HttpStatusCode.Ambiguous;
+                Guid tempGuid = _fileManager.Save(fileName, fileBytes);
+                myResponse.Content = new StringContent(tempGuid.ToString());
+                myResponse.Content.Headers.ContentType = new MediaTypeHeaderValue("text/plain");
 
                 return myResponse;
             }
@@ -185,15 +211,22 @@ namespace ServerFileSync.Controllers
         /// <param name="extension"></param>
         /// <returns></returns>
         [HttpDelete]
-        public HttpResponseMessage Delete(string filename)
+        public HttpResponseMessage Delete(string filename, string previousHash)
         {
             try
             {
                 if (String.IsNullOrEmpty(filename))
                     return new HttpResponseMessage(HttpStatusCode.BadRequest);
-                _fileManager.Delete(filename);
-                _hubWrapper.NotifyDeleteFile(filename);
-                return new HttpResponseMessage(HttpStatusCode.OK);
+
+                if (String.IsNullOrEmpty(previousHash)||(_fileManager.GetHash(filename).Equals(previousHash)))
+                {
+                    _fileManager.Delete(filename);
+                    _hubWrapper.NotifyDeleteFile(filename);
+                    return new HttpResponseMessage(HttpStatusCode.OK);
+                }
+
+                //File intended to Delete was modified. Must confirm Delete calling it with previousHash empty.
+                return new HttpResponseMessage(HttpStatusCode.Ambiguous);
             }
             catch (IOException excp)
             {
@@ -209,21 +242,29 @@ namespace ServerFileSync.Controllers
         /// Checks the existance of a file
         /// </summary>
         /// <param name="fileName"></param>
-        /// <returns>true if exists, false if it doesn't</returns>
+        /// <returns>
+        /// String Content: true if name exists, false if it doesn't
+        ///// HttpStatusCode:
+        ///// Ok: if file exists with given name and hash,
+        ///// Ambiguous: if file exists with given name but different hash.
+        /// </returns>
         [HttpGet]
-        public HttpResponseMessage Exists(string fileName)
+        public HttpResponseMessage Exists(string fileName/*, string hash*/)
         {
-            var response = new HttpResponseMessage(HttpStatusCode.OK);
             if (String.IsNullOrEmpty(fileName))
-            {
-                response.StatusCode = HttpStatusCode.BadRequest;
-                return response;
-            }
-            else
-            {
-                response.Content = new StringContent(_fileManager.Exists(fileName) ? "true" : "false");
-                return response;
-            }
+                return new HttpResponseMessage(HttpStatusCode.BadRequest);
+
+            bool exists = _fileManager.Exists(fileName);
+            var response = new HttpResponseMessage(HttpStatusCode.OK);
+            response.Content = new StringContent( exists? "true" : "false");
+
+            /*
+            //if file exists but has different hash
+            if (exists && !_fileManager.GetHash(fileName).Equals(hash))
+                response.StatusCode = HttpStatusCode.Ambiguous;
+            */
+
+            return response;
         }
 
         /// <summary>
@@ -243,16 +284,40 @@ namespace ServerFileSync.Controllers
             return fullResponse;
         }
 
+        [HttpGet]
+        public HttpResponseMessage GetFolderStatus()
+        {
+            List<FolderFileState> remoteFiles = new List<FolderFileState>();
+            var files = _fileManager.GetFilenames();
+
+            foreach (string file in files)
+                remoteFiles.Add(new FolderFileState()
+                        .FileName(file)
+                        .Hash(_fileManager.GetHash(file))
+                        .CurrentStatus(FileStatusType.Synced));
+
+            FolderState folderState = new FolderState()
+                .RemotePath(_syncFolder);
+
+            foreach (var f in remoteFiles)
+                folderState.RemoteFile(f);
+
+            string s = folderState.Definition;
+
+            var resp = new HttpResponseMessage { Content = new StringContent(s, System.Text.Encoding.UTF8, "application/xml") };
+            return resp;
+        }
+
         #endregion Public methods
 
         #region Private methods
         private async Task<byte[]> getFileBytesFromRequest(MultipartMemoryStreamProvider multiContents)
         {
             byte[] fileBytes;
-            if (multiContents.Contents.Count > 1)
+            if (multiContents.Contents.Count > 2)
             {
                 //The file as byte array
-                fileBytes = await multiContents.Contents[1].ReadAsByteArrayAsync();
+                fileBytes = await multiContents.Contents[2].ReadAsByteArrayAsync();
                 if (fileBytes.Length <= 0)
                     throw new HttpResponseException(HttpStatusCode.BadRequest);
             }
